+++ server/dao/NewsletterDAO.js
@@ -1,0 +1,4 @@
+const mongoose = require("mongoose")
+const NewsletterSchema = require("../models/Newsletter")
+
+module.exports = mongoose.model('Newsletter', NewsletterSchema)
