+++ server/package.json
@@ -11,6 +11,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "agenda": "^4.1.0",
     "bcrypt": "^5.0.0",
     "body-parser": "^1.19.0",
     "demo-project": "^2.0.0",
