+++ server/server.js
@@ -8,13 +8,14 @@ const db = require("./config/database");
 const properties = require("./config/properties");
 const clientRouter = require("./routes/clientrouter.js");
 const crudRouter = require("./routes/crud");
-
+const scheduler = require("./config/scheduler")
 
 app.use(express.static(path.join(__dirname, '../build')));
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 app.use(morgan('combined'));
 app.use(actuator());
+
 app.use(properties.database_ENDPOINT, crudRouter);
 app.use(properties.client_ENDPOINT, clientRouter);
 
@@ -23,6 +24,7 @@ app.get('/*', function(req, res){
 });
 
 db();
+scheduler.start();
 
 app.listen(properties.SERVER_PORT, function () {
   console.log(`Listening on port: ${properties.SERVER_PORT}`);
